+++ js/main.js
@@ -1,6 +1,6 @@
 import {renderPhotoUsers} from './renderingThumbnails.js';
 import {initBigPicture} from './renderingBigPicture.js';
-import {validatePictureForm} from './newPictureForm.js';
+import {newPictureForm} from './newPictureForm.js';
 renderPhotoUsers();
 initBigPicture();
-validatePictureForm();
+newPictureForm();
